feat(client): privy login
- polyfill buffer because of vite
- add process.env because of vite (and privy's react lib uses it)
