+++ packages/client/package.json
@@ -28,6 +28,7 @@
     "viem": "1.6.0"
   },
   "devDependencies": {
+    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
     "@types/react": "18.2.22",
     "@types/react-dom": "18.2.7",
     "@vitejs/plugin-react": "^3.1.0",
