+++ packages/client/src/App.tsx
@@ -1,29 +1,21 @@
-import { useComponentValue } from "@latticexyz/react";
-import { useMUD } from "./MUDContext";
-import { singletonEntity } from "@latticexyz/store-sync/recs";
+import { PrivyProvider } from "@privy-io/react-auth";
+import Main from "./components/Main";
 
-export const App = () => {
-  const {
-    components: { Counter },
-    systemCalls: { increment },
-  } = useMUD();
+const PRIVY_APP_ID = import.meta.env.VITE_PRIVY_APP_ID;
 
-  const counter = useComponentValue(Counter, singletonEntity);
+const App = () => {
+  if (!PRIVY_APP_ID) throw new Error("[App] Missing Privy app id");
 
   return (
-    <>
-      <div>
-        Counter: <span>{counter?.value ?? "??"}</span>
-      </div>
-      <button
-        type="button"
-        onClick={async (event) => {
-          event.preventDefault();
-          console.log("new counter value:", await increment());
-        }}
-      >
-        Increment
-      </button>
-    </>
+    <PrivyProvider
+      appId={PRIVY_APP_ID}
+      config={{
+        loginMethods: ["email", "wallet"],
+      }}
+    >
+      <Main />
+    </PrivyProvider>
   );
 };
+
+export default App;
