+++ packages/client/src/index.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from "react-dom/client";
-import { App } from "./App";
+import App from "./App";
 import { setup } from "./mud/setup";
 import { MUDProvider } from "./MUDContext";
 import mudConfig from "contracts/mud.config";
@@ -8,7 +8,6 @@ const rootElement = document.getElementById("react-root");
 if (!rootElement) throw new Error("React root not found");
 const root = ReactDOM.createRoot(rootElement);
 
-// TODO: figure out if we actually want this to be async or if we should render something else in the meantime
 setup().then(async (result) => {
   root.render(
     <MUDProvider value={result}>
